add updated screenshots, fix clippy warnings

+++ src/clip.rs
@@ -50,11 +50,11 @@ pub fn get_time_range(marker: &Marker, max_duration: Option<u32>) -> (u32, Optio
         .iter()
         .find(|m| m.start > marker.start);
     if let Some(max_duration) = max_duration {
-        (start as u32, Some(start as u32 + max_duration))
+        (start, Some(start + max_duration))
     } else if let Some(next) = next_marker {
-        (start as u32, Some(next.start as u32))
+        (start, Some(next.start))
     } else {
-        (start as u32, None)
+        (start, None)
     }
 }
